backend/pyproject.toml
[project]
name = "trajectory-representation-visual-analytics-backend"
version = "0.1.0"
description = "FastAPI backend for trajectory representation visual analytics"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "numpy>=1.26",
  "scikit-learn>=1.4",
  "umap-learn>=0.5.6",
  "sude",
  "ipykernel>=7.2.0",
]

[tool.uv]
package = false
